Refactor.
* Clean up published interface.
* Move internal stuff to internal package.
* Move test utilities to util package.
* Cli test.
* exclude debug stuff from coverage.
* make imports consistent
* remove some needless wildcard imports
* run intellij formatter on all project (mostly to organize imports).
  This thing can break code!
* Rename a few modules.
