+++ core/src/main/scala/org/scalafmt/ScalaStyle.scala
@@ -3,28 +3,21 @@ package org.scalafmt
 import scala.concurrent.duration.Duration
 
 /**
-  * A scalafmt style.
-  *
-  * To support pickling with upickle, it's defined as a trait. It could be a
-  * sealed abstract class.
+  * A scalafmt style, use [[ScalaStyle.Standard]].
   */
+// TODO(olafur) refactor to case class or something simpler
 sealed trait ScalaStyle {
-  // TODO(olafur) soft column-limit?
   /**
     * Column limit, any formatting exceeding this field is penalized heavily.
     */
+  // TODO(olafur) soft column-limit?
   def maxColumn: Int = 80
-
-  /**
-    * See [[ArgumentHandling]].
-    */
-  def argumentHandling: ArgumentHandling = OneArgOneLine
 
   /**
     * The maximum duration the formatter is allowed to run before giving up.
     * If the formatter times out, the original source code is returned.
     */
-  def maxDuration: Duration = Duration(200, "ms")
+  def maxDuration: Duration = Duration(400, "ms")
 
   /**
     * Debugging only. Maximum number of states that the formatter may visit.
@@ -41,26 +34,34 @@ sealed trait ScalaStyle {
 }
 
 /**
-  * Recommended style if you are not sure which one to pick.
-  */
-case object Standard extends ScalaStyle
-
-/**
   * Debugging only. Used in unit tests.
   */
-sealed trait UnitTestStyle extends ScalaStyle {
+protected[scalafmt] sealed trait UnitTestStyle extends ScalaStyle {
+  override val debug = true
+
   override def maxDuration = Duration(10, "s")
-  override val debug = true
 }
 
-case object ManualTestStyle extends ScalaStyle {
-  override def maxStateVisits = 20000
-  override def maxDuration = Duration(10, "min")
-  override val debug = true
+object ScalaStyle {
+
+  /**
+    * Recommended style if you are not sure which one to pick.
+    */
+  case object Standard extends ScalaStyle
+
+  protected[scalafmt] case object ManualTest extends ScalaStyle {
+    override val debug = true
+
+    override def maxStateVisits = 20000
+
+    override def maxDuration = Duration(10, "min")
+  }
+
+  protected[scalafmt] case object UnitTest80 extends UnitTestStyle
+
+  protected[scalafmt] case object UnitTest40 extends UnitTestStyle {
+    override def maxColumn = 40
+  }
+
 }
 
-case object UnitTest80 extends UnitTestStyle
-
-case object UnitTest40 extends UnitTestStyle {
-  override def maxColumn = 40
-}
