+++ core/src/main/scala/org/scalafmt/internal/ScalaFmtLogger.scala
@@ -1,8 +1,8 @@
-package org.scalafmt
+package org.scalafmt.internal
 
-import java.util.concurrent.TimeUnit
 import com.typesafe.scalalogging.Logger
 import org.slf4j.LoggerFactory
+
 import scala.meta.Tree
 import scala.meta.prettyprinters.Structure
 import scala.meta.tokens.Token
@@ -15,26 +15,26 @@ trait ScalaFmtLogger {
 
   def log(formatToken: FormatToken): String =
     s"""${log(formatToken.left)}
-       |${log(formatToken.between:_*)}
+       |${log(formatToken.between: _*)}
        |${log(formatToken.right)}""".stripMargin
 
-  def log(tokens: Token *): String = tokens.map(log).mkString("\n")
+  def log(tokens: Token*): String = tokens.map(log).mkString("\n")
+
+  def small(token: Token): String =
+    token.code.replace("\n", "").slice(0, 10)
+
+  def log(tokens: Tokens): String = tokens.map(log).mkString("\n")
 
   def log(token: Token): String = f"$token%-30s ${getTokenClass(token)}"
 
-  def small(token: Token): String =
-    token.code.replace("\n", "").slice(0,10)
-
   private def getTokenClass(token: Token) =
     token.getClass.getName.stripPrefix("scala.meta.tokens.Token$")
-
-  def log(tokens: Tokens): String = tokens.map(log).mkString("\n")
 
   def log(t: Tree): String = {
     s"""TYPE: ${t.getClass.getName.stripPrefix("scala.meta.")}
         |SOURCE: $t
         |STRUCTURE: ${t.show[Structure]}
-        |TOKENS: ${t.tokens.map(x=>reveal(x.code)).mkString(",")}
+        |TOKENS: ${t.tokens.map(x => reveal(x.code)).mkString(",")}
         |""".stripMargin
   }
 
@@ -45,7 +45,7 @@ trait ScalaFmtLogger {
       case ch => ch
     }
 
-  def header [T](t: T): String = {
+  def header[T](t: T): String = {
     val line = s"=" * (t.toString.length + 3)
     s"$line\n=> $t\n$line"
   }
