+++ core/src/main/scala/org/scalafmt/internal/Split.scala
@@ -1,4 +1,4 @@
-package org.scalafmt
+package org.scalafmt.internal
 
 import scala.meta.tokens.Token
 
@@ -13,15 +13,20 @@ import scala.meta.tokens.Token
   * @param indents Does this add indentation?
   * @param policy How does this split affect other later splits?
   * @param penalty Does this split overflow the column limit?
-  * @param line For debugging, to retrace from which case in [[Formatter]]
+  * @param line For debugging, to retrace from which case in [[Router]]
   *             this split originates.
   *
   */
 case class Split(modification: Modification, cost: Int, ignoreIf: Boolean =
-    false, indents: List[Indent[Length]] = List.empty[Indent[Length]],
-    policy: Policy = NoPolicy, penalty: Boolean =
-    false,
-    optimalAt: Option[Token] = None)(implicit val line: sourcecode.Line) {
+false, indents: List[Indent[Length]] = List.empty[Indent[Length]],
+                 policy: Policy = NoPolicy, penalty: Boolean =
+                 false,
+                 optimalAt: Option[Token] = None)(implicit val line: sourcecode.Line) {
+
+  val indentation = indents.map(_.length match {
+    case Num(x) => x.toString
+    case x => x.toString
+  }).mkString("[", ", ", "]")
 
   def length: Int =
     modification match {
@@ -30,60 +35,55 @@ case class Split(modification: Modification, cost: Int, ignoreIf: Boolean =
       case Space => 1
       case Provided(code) =>
         val firstLine = code.indexOf("\n")
-        if (firstLine == - 1) code.length
+        if (firstLine == -1) code.length
         else firstLine
     }
 
   def withOptimal(token: Token): Split =
     new Split(modification, cost, ignoreIf, indents, policy, true, Some(token))(
-        line)
+      line)
 
   def withPolicy(newPolicy: Policy): Split = {
     val update =
       if (policy == NoPolicy) newPolicy
       else newPolicy orElse policy
     new Split(modification, cost, ignoreIf, indents, update, true, optimalAt)(
-        line)
+      line)
   }
 
   def withPenalty(penalty: Int): Split =
     new Split(modification,
-              cost + penalty,
-              ignoreIf,
-              indents,
-              policy,
-              true,
-              optimalAt)(line)
+      cost + penalty,
+      ignoreIf,
+      indents,
+      policy,
+      true,
+      optimalAt)(line)
 
   def withIndent(length: Length, expire: Token,
-          expiresOn: ExpiresOn): Split = {
+                 expiresOn: ExpiresOn): Split = {
     length match {
       case Num(0) => this
       case _ =>
         new Split(modification,
-                  cost,
-                  ignoreIf,
-                  Indent(length, expire, expiresOn) +: indents,
-                  policy,
-                  penalty,
-                  optimalAt)(line)
+          cost,
+          ignoreIf,
+          Indent(length, expire, expiresOn) +: indents,
+          policy,
+          penalty,
+          optimalAt)(line)
     }
   }
-  val indentation = indents.map(_.length match {
-        case Num(x) => x.toString
-        case x => x.toString
-      }).mkString("[", ", ", "]")
 
   def withModification(newModification: Modification): Split =
     new Split(
-        newModification, cost, ignoreIf, indents, policy, penalty, optimalAt)(
-        line)
-
-
-  def hasPolicy = policy ne NoPolicy
+      newModification, cost, ignoreIf, indents, policy, penalty, optimalAt)(
+      line)
 
   override def toString =
     s"""$modification:${line.value}(cost=$cost, indents=$indentation, p=$hasPolicy)"""
+
+  def hasPolicy = policy ne NoPolicy
 
   // TODO(olafur) come with better concept of split equality.
