+++ core/src/main/scala/org/scalafmt/internal/package.scala
@@ -1,22 +1,24 @@
-package org
+package org.scalafmt
 
 import scala.annotation.tailrec
 import scala.meta.Tree
 import scala.meta.tokens.Token
 
-package object scalafmt {
+package object internal {
   type Policy = PartialFunction[Decision, Decision]
   val NoPolicy = PartialFunction.empty[Decision, Decision]
-  val IdentityPolicy: PartialFunction[Decision, Decision] = { case d => d }
+  val IdentityPolicy: PartialFunction[Decision, Decision] = {
+    case d => d
+  }
 
   // TODO(olafur) Move these elsewhere.
 
   @tailrec
   final def childOf(child: Tree, tree: Tree): Boolean = {
     child == tree || (child.parent match {
-        case Some(parent) => childOf(parent, tree)
-        case _ => false
-      })
+      case Some(parent) => childOf(parent, tree)
+      case _ => false
+    })
   }
 
   def childOf(tok: Token, tree: Tree, owners: Map[Token, Tree]): Boolean =
