+++ core/src/test/scala/org/scalafmt/util/ErrorTest.scala
@@ -1,0 +1,19 @@
+package org.scalafmt.util
+
+import org.scalafmt.ScalaFmt
+import org.scalafmt.ScalaStyle
+import org.scalatest.FunSuite
+
+class ErrorTest extends FunSuite {
+  test("errors are caught") {
+    val nonSourceFile = Seq(
+      "class A {",
+      "val x = 1",
+      "println(1)"
+    )
+    nonSourceFile.foreach { original =>
+      val obtained = ScalaFmt.format(original, ScalaStyle.UnitTest40)
+      DiffUtil.assertNoDiff(obtained, original)
+    }
+  }
+}
