+++ core/src/test/scala/org/scalafmt/util/FormatOutput.scala
@@ -1,3 +1,3 @@
-package org.scalafmt
+package org.scalafmt.util
 
 case class FormatOutput(token: String, whitespace: String, visits: Int)
