+++ core/src/test/scala/org/scalafmt/util/Tabulator.scala
@@ -1,4 +1,4 @@
-package org.scalafmt
+package org.scalafmt.util
 
 /**
   * Shamelessly copied from:
@@ -11,27 +11,29 @@ object Tabulator {
       case Seq() => ""
       case _ =>
         val sizes = for (
-        row <- table) yield ( for (
-        cell <- row) yield if (cell == null) 0
+          row <- table) yield (for (
+          cell <- row) yield if (cell == null) 0
         else cell.toString.length)
         val colSizes = for (
-        col <- sizes.transpose) yield col.max
+          col <- sizes.transpose) yield col.max
         val rows = for (
-        row <- table) yield formatRow(row, colSizes)
+          row <- table) yield formatRow(row, colSizes)
         formatRows(rowSeparator(colSizes), rows)
     }
 
   def formatRows(rowSeparator: String, rows: Seq[String]): String =
     (rowSeparator :: rows.head :: rowSeparator :: rows.tail.toList ::: rowSeparator :: List(
-          )).mkString("\n")
+    )).mkString("\n")
 
   def formatRow(row: Seq[Any], colSizes: Seq[Int]) = {
-    val cells = ( for (
-    (item, size) <- row.zip(colSizes)) yield if (size == 0) ""
+    val cells = (for (
+      (item, size) <- row.zip(colSizes)) yield if (size == 0) ""
     else ("%" + size + "s").format(item))
     cells.mkString("|", "|", "|")
   }
 
   def rowSeparator(colSizes: Seq[Int]) =
-    colSizes map { "-" * _ } mkString ("+", "+", "+")
+    colSizes map {
+      "-" * _
+    } mkString("+", "+", "+")
 }
